+++ legacy/evas/src/lib/engines/common_16/evas_soft16_main.c
@@ -6,21 +6,21 @@ static int               reference = 0;
 static Image_Entry      *_evas_common_soft16_image_new(void);
 static void              _evas_common_soft16_image_delete(Image_Entry *ie);
 
-static int               _evas_common_soft16_image_surface_alloc(Image_Entry *ie, int w, int h);
+static int               _evas_common_soft16_image_surface_alloc(Image_Entry *ie, unsigned int w, unsigned int h);
 static void              _evas_common_soft16_image_surface_delete(Image_Entry *ie);
 static DATA32		*_evas_common_soft16_image_surface_pixels(Image_Entry *ie);
 
 static int               _evas_common_load_soft16_image_from_file(Image_Entry *ie);
 static void              _evas_common_soft16_image_unload(Image_Entry *ie);
 
-static void              _evas_common_soft16_image_dirty_region(Image_Entry *im, int x, int y, int w, int h);
+static void              _evas_common_soft16_image_dirty_region(Image_Entry *im, unsigned int x, unsigned int y, unsigned int w, unsigned int h);
 static int               _evas_common_soft16_image_dirty(Image_Entry *ie_dst, const Image_Entry *ie_src);
 
 static int               _evas_common_soft16_image_ram_usage(Image_Entry *ie);
 
-static int               _evas_common_soft16_image_size_set(Image_Entry *ie_dst, const Image_Entry *ie_im, int w, int h);
-static int               _evas_common_soft16_image_from_copied_data(Image_Entry* ie_dst, int w, int h, DATA32 *image_data, int alpha, int cspace);
-static int               _evas_common_soft16_image_from_data(Image_Entry* ie_dst, int w, int h, DATA32 *image_data, int alpha, int cspace);
+static int               _evas_common_soft16_image_size_set(Image_Entry *ie_dst, const Image_Entry *ie_im, unsigned int w, unsigned int h);
+static int               _evas_common_soft16_image_from_copied_data(Image_Entry* ie_dst, unsigned int w, unsigned int h, DATA32 *image_data, int alpha, int cspace);
+static int               _evas_common_soft16_image_from_data(Image_Entry* ie_dst, unsigned int w, unsigned int h, DATA32 *image_data, int alpha, int cspace);
 static int               _evas_common_soft16_image_colorspace_set(Image_Entry* ie_dst, int cspace);
 
 static int               _evas_common_load_soft16_image_data_from_file(Image_Entry *ie);
@@ -112,7 +112,7 @@ _evas_common_soft16_image_delete(Image_Entry *ie)
 }
 
 static int
-_evas_common_soft16_image_surface_alloc(Image_Entry *ie, int w, int h)
+_evas_common_soft16_image_surface_alloc(Image_Entry *ie, unsigned int w, unsigned int h)
 {
    Soft16_Image *im = (Soft16_Image *) ie;
 
@@ -181,7 +181,7 @@ _evas_common_soft16_image_unload(Image_Entry *ie __UNUSED__)
 }
 
 static void
-_evas_common_soft16_image_dirty_region(Image_Entry *im __UNUSED__, int x __UNUSED__, int y __UNUSED__, int w __UNUSED__, int h __UNUSED__)
+_evas_common_soft16_image_dirty_region(Image_Entry *im __UNUSED__, unsigned int x __UNUSED__, unsigned int y __UNUSED__, unsigned int w __UNUSED__, unsigned int h __UNUSED__)
 {
 }
 
@@ -211,7 +211,7 @@ _evas_common_soft16_image_ram_usage(Image_Entry *ie)
 }
 
 static int
-_evas_common_soft16_image_size_set(Image_Entry *ie_dst, const Image_Entry *ie_im, int w __UNUSED__, int h __UNUSED__)
+_evas_common_soft16_image_size_set(Image_Entry *ie_dst, const Image_Entry *ie_im, unsigned int w __UNUSED__, unsigned int h __UNUSED__)
 {
    Soft16_Image *dst = (Soft16_Image *) ie_dst;
    Soft16_Image *im = (Soft16_Image *) ie_im;
@@ -222,7 +222,7 @@ _evas_common_soft16_image_size_set(Image_Entry *ie_dst, const Image_Entry *ie_im
 }
 
 static int
-_evas_common_soft16_image_from_data(Image_Entry* ie_dst, int w, int h, DATA32 *image_data, int alpha, int cspace __UNUSED__)
+_evas_common_soft16_image_from_data(Image_Entry* ie_dst, unsigned int w, unsigned int h, DATA32 *image_data, int alpha, int cspace __UNUSED__)
 {
    Soft16_Image *im = (Soft16_Image *) ie_dst;
 
@@ -245,7 +245,7 @@ _evas_common_soft16_image_from_data(Image_Entry* ie_dst, int w, int h, DATA32 *i
 }
 
 static int
-_evas_common_soft16_image_from_copied_data(Image_Entry* ie_dst, int w __UNUSED__, int h, DATA32 *image_data, int alpha __UNUSED__, int cspace __UNUSED__)
+_evas_common_soft16_image_from_copied_data(Image_Entry* ie_dst, unsigned int w __UNUSED__, unsigned int h, DATA32 *image_data, int alpha __UNUSED__, int cspace __UNUSED__)
 {
    Soft16_Image *im = (Soft16_Image *) ie_dst;
 
@@ -337,8 +337,8 @@ _get_clip(const RGBA_Draw_Context *dc, const Soft16_Image *im,
 	     clip->h += clip->y;
 	     clip->y = 0;
 	  }
-	if ((clip->x + clip->w) > im->cache_entry.w) clip->w = im->cache_entry.w - clip->x;
-	if ((clip->y + clip->h) > im->cache_entry.h) clip->h = im->cache_entry.h - clip->y;
+	if ((clip->x + clip->w) > (int)im->cache_entry.w) clip->w = im->cache_entry.w - clip->x;
+	if ((clip->y + clip->h) > (int)im->cache_entry.h) clip->h = im->cache_entry.h - clip->y;
      }
    else
      {
