+++ legacy/evas/src/lib/engines/common_8/evas_soft8_main.c
@@ -6,31 +6,23 @@ static int reference = 0;
 static Image_Entry *_evas_common_soft8_image_new(void);
 static void _evas_common_soft8_image_delete(Image_Entry * ie);
 
-static int _evas_common_soft8_image_surface_alloc(Image_Entry * ie, int w,
-                                                  int h);
+static int _evas_common_soft8_image_surface_alloc(Image_Entry * ie, unsigned int w, unsigned int h);
 static void _evas_common_soft8_image_surface_delete(Image_Entry * ie);
 static DATA32 *_evas_common_soft8_image_surface_pixels(Image_Entry * ie);
 
 static int _evas_common_load_soft8_image_from_file(Image_Entry * ie);
 static void _evas_common_soft8_image_unload(Image_Entry * ie);
 
-static void _evas_common_soft8_image_dirty_region(Image_Entry * im, int x,
-                                                  int y, int w, int h);
+static void _evas_common_soft8_image_dirty_region(Image_Entry * im, unsigned int x, unsigned int y, unsigned int w, unsigned int h);
 static int _evas_common_soft8_image_dirty(Image_Entry * ie_dst,
                                           const Image_Entry * ie_src);
 
 static int _evas_common_soft8_image_ram_usage(Image_Entry * ie);
 
 static int _evas_common_soft8_image_size_set(Image_Entry * ie_dst,
-                                             const Image_Entry * ie_im, int w,
-                                             int h);
-static int _evas_common_soft8_image_from_copied_data(Image_Entry * ie_dst,
-                                                     int w, int h,
-                                                     DATA32 * image_data,
-                                                     int alpha, int cspace);
-static int _evas_common_soft8_image_from_data(Image_Entry * ie_dst, int w,
-                                              int h, DATA32 * image_data,
-                                              int alpha, int cspace);
+                                             const Image_Entry * ie_im, unsigned int w, unsigned int h);
+static int _evas_common_soft8_image_from_copied_data(Image_Entry * ie_dst, unsigned int w, unsigned int h, DATA32 * image_data, int alpha, int cspace);
+static int _evas_common_soft8_image_from_data(Image_Entry * ie_dst, unsigned int w, unsigned int h, DATA32 * image_data, int alpha, int cspace);
 static int _evas_common_soft8_image_colorspace_set(Image_Entry * ie_dst,
                                                    int cspace);
 
@@ -125,7 +117,7 @@ _evas_common_soft8_image_delete(Image_Entry * ie)
 }
 
 static int
-_evas_common_soft8_image_surface_alloc(Image_Entry * ie, int w, int h)
+_evas_common_soft8_image_surface_alloc(Image_Entry * ie, unsigned int w, unsigned int h)
 {
    Soft8_Image *im = (Soft8_Image *) ie;
 
@@ -203,8 +195,10 @@ _evas_common_soft8_image_unload(Image_Entry * ie __UNUSED__)
 
 static void
 _evas_common_soft8_image_dirty_region(Image_Entry * im __UNUSED__,
-                                      int x __UNUSED__, int y __UNUSED__,
-                                      int w __UNUSED__, int h __UNUSED__)
+                                      unsigned int x __UNUSED__,
+                                      unsigned int y __UNUSED__,
+                                      unsigned int w __UNUSED__,
+                                      unsigned int h __UNUSED__)
 {
 }
 
@@ -235,8 +229,9 @@ _evas_common_soft8_image_ram_usage(Image_Entry * ie)
 
 static int
 _evas_common_soft8_image_size_set(Image_Entry * ie_dst,
-                                  const Image_Entry * ie_im, int w __UNUSED__,
-                                  int h __UNUSED__)
+                                  const Image_Entry * ie_im,
+                                  unsigned int w __UNUSED__,
+                                  unsigned int h __UNUSED__)
 {
    Soft8_Image *dst = (Soft8_Image *) ie_dst;
    Soft8_Image *im = (Soft8_Image *) ie_im;
@@ -247,7 +242,8 @@ _evas_common_soft8_image_size_set(Image_Entry * ie_dst,
 }
 
 static int
-_evas_common_soft8_image_from_data(Image_Entry * ie_dst, int w, int h,
+_evas_common_soft8_image_from_data(Image_Entry * ie_dst,
+                                   unsigned int w, unsigned int h,
                                    DATA32 * image_data, int alpha,
                                    int cspace __UNUSED__)
 {
@@ -273,7 +269,8 @@ _evas_common_soft8_image_from_data(Image_Entry * ie_dst, int w, int h,
 
 static int
 _evas_common_soft8_image_from_copied_data(Image_Entry * ie_dst,
-                                          int w __UNUSED__, int h,
+                                          unsigned int w __UNUSED__,
+                                          unsigned int h,
                                           DATA32 * image_data,
                                           int alpha __UNUSED__,
                                           int cspace __UNUSED__)
@@ -330,7 +327,7 @@ _evas_common_load_soft8_image_data_from_file(Image_Entry * ie)
 }
 
 /* Soft16_Image * */
-/* evas_common_soft16_image_new(int w, int h, int stride, int have_alpha, DATA16 *pixels, */
+/* evas_common_soft16_image_new(int w, unsigned int h, unsigned int stride, int have_alpha, DATA16 *pixels, */
 /* 		 int copy) */
 /* { */
 /*    Soft16_Image *im; */
@@ -371,9 +368,9 @@ _get_clip(const RGBA_Draw_Context * dc, const Soft8_Image * im,
              clip->h += clip->y;
              clip->y = 0;
           }
-        if ((clip->x + clip->w) > im->cache_entry.w)
+        if ((clip->x + clip->w) > (int)im->cache_entry.w)
            clip->w = im->cache_entry.w - clip->x;
-        if ((clip->y + clip->h) > im->cache_entry.h)
+        if ((clip->y + clip->h) > (int)im->cache_entry.h)
            clip->h = im->cache_entry.h - clip->y;
      }
    else
@@ -407,8 +404,7 @@ _shrink(int *s_pos, int *s_size, int pos, int size)
 
 static int
 _soft8_adjust_areas(Eina_Rectangle * src,
-                    int src_max_x, int src_max_y,
-                    Eina_Rectangle * dst,
+                    int src_max_x, int src_max_y, Eina_Rectangle * dst,
                     int dst_max_x, int dst_max_y, Eina_Rectangle * dst_clip)
 {
    if (_is_empty_rectangle(src) ||
@@ -524,7 +520,8 @@ evas_common_soft8_image_draw(Soft8_Image * src, Soft8_Image * dst,
                  int src_region_x, int src_region_y,
                  int src_region_w, int src_region_h,
                  int dst_region_x, int dst_region_y,
-                 int dst_region_w, int dst_region_h, int smooth __UNUSED__)
+                 int dst_region_w, int dst_region_h,
+                 int smooth __UNUSED__)
 {
    Eina_Rectangle sr, dr;
    Cutout_Rects *rects;
@@ -605,7 +602,7 @@ evas_common_soft8_image_alpha_set(Soft8_Image * im, int have_alpha)
 }
 
 /* Soft16_Image * */
-/* evas_common_soft16_image_size_set(Soft16_Image *old_im, int w, int h) */
+/* evas_common_soft16_image_size_set(Soft16_Image *old_im, unsigned int w, unsigned int h) */
 /* { */
 /*    Soft16_Image *new_im; */
 /*    DATA16 *dp, *sp; */
