+++ legacy/evas/src/modules/engines/software_8_x11/evas_engine.c
@@ -78,13 +78,13 @@ static void
 _tmp_out_alloc(Render_Engine * re)
 {
    Tilebuf_Rect *r;
-   int w = 0, h = 0;
+   unsigned int w = 0, h = 0;
 
    EINA_INLIST_FOREACH(re->rects, r)
    {
-      if (r->w > w)
+      if (r->w > (int)w)
          w = r->w;
-      if (r->h > h)
+      if (r->h > (int)h)
          h = r->h;
    }
 
@@ -512,8 +512,9 @@ _tmp_out_process(Render_Engine * re, int out_x, int out_y, int w, int h)
    d = re->shbuf->im;
    s = re->tmp_out;
 
-   if ((w < 1) || (h < 1) || (out_x >= d->cache_entry.w)
-       || (out_y >= d->cache_entry.h))
+   if ((w < 1) || (h < 1) ||
+       (out_x >= (int)d->cache_entry.w) ||
+       (out_y >= (int)d->cache_entry.h))
       return;
 
    if (re->rot == 90)
