+++ legacy/evas/src/modules/engines/software_generic/evas_engine.c
@@ -520,12 +520,12 @@ eng_image_map4_draw(void *data __UNUSED__, void *context, void *surface, void *i
        (p[0].y <= p[2].y) &&
        (p[0].u == 0) &&
        (p[0].v == 0) &&
-       (p[1].u == (im->cache_entry.w << FP)) &&
+       (p[1].u == (int)(im->cache_entry.w << FP)) &&
        (p[1].v == 0) &&
-       (p[2].u == (im->cache_entry.w << FP)) &&
-       (p[2].v == (im->cache_entry.h << FP)) &&
+       (p[2].u == (int)(im->cache_entry.w << FP)) &&
+       (p[2].v == (int)(im->cache_entry.h << FP)) &&
        (p[3].u == 0) &&
-       (p[3].v == (im->cache_entry.h << FP)) &&
+       (p[3].v == (int)(im->cache_entry.h << FP)) &&
        (p[0].col == 0xffffffff) &&
        (p[1].col == 0xffffffff) &&
        (p[2].col == 0xffffffff) &&
